changed name to setToggle
