+++ src/store/modules/app/state.js
@@ -1,5 +1,5 @@
 export default {
-  drawer: null,
+  drawer: null, //controls the side bar "SideDrawer" drawer open/close state
   color: 'general',
   image: 'https://demos.creative-tim.com/vue-material-dashboard/img/sidebar-3.3a54f533.jpg' //change if you want a picture but if you are using this internally get rid of the picture links here and in filter.vue
 }
